@@ -1,6 +1,6 @@
 {
   "name": "@nieuwlandgeo/sldreader",
-  "version": "0.0.6",
+  "version": "0.2.6",
   "description": "SLD reader and formatter for openlayers",
   "main": "dist/sldreader.js",
   "keywords": [
@@ -9,12 +9,11 @@
   "scripts": {
     "build": "rollup -c",
     "docker": "docker-compose up -d",
-    "docs:build": "documentation build src/index.js -c docs/documentation.yml -f md --output docs/apigen.md",
+    "docs:build": "jsdoc2md 'src/**/*.js' > docs/apigen.md",
     "docs": "npm-run-all build docs:build && cp dist/sldreader.js docs/assets/sldreader.js",
     "lint": "eslint src/ test/",
     "lint:win": "eslint src/ test/ --rule \"linebreak-style: off\"",
     "postversion": "git push --tags && npm publish && git rm -r dist && git commit -m 'remove dist' && git push",
-    "preversion": "npm test",
     "start": "npm-run-all -p start:* docker",
     "start:watch": "rollup -c -w --file docs/assets/sldreader.js",
     "start:sync": "browser-sync start --config bs-config.js",
@@ -33,34 +32,36 @@
   "author": "Allart Kooiman <akooiman@nieuwland.nl>",
   "license": "ISC",
   "devDependencies": {
+    "@rollup/plugin-buble": "^0.21.3",
+    "@rollup/plugin-node-resolve": "^7.1.3",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-latest": "^6.24.1",
-    "browser-sync": "^2.26.3",
+    "browser-sync": "^2.26.7",
     "chai": "^4.2.0",
-    "core-js": "^2.6.5",
-    "documentation": "^9.3.1",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.16.0",
-    "karma": "^3.1.4",
+    "core-js": "^3.6.5",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb-base": "^14.1.0",
+    "eslint-plugin-import": "^2.20.2",
+    "jsdoc-to-markdown": "^5.0.3",
+    "karma": "^5.0.2",
     "karma-chai": "^0.1.0",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-mocha": "^1.3.0",
+    "karma-firefox-launcher": "^1.3.0",
+    "karma-mocha": "^2.0.0",
     "karma-mocha-reporter": "^2.2.5",
-    "karma-rollup-preprocessor": "^7.0.0",
-    "mocha": "^5.2.0",
+    "karma-rollup-preprocessor": "^7.0.5",
+    "mocha": "^7.1.1",
     "npm-run-all": "^4.1.5",
-    "ol": "^4.6.5",
-    "prettier": "^1.16.4",
-    "rollup": "^1.7.4",
-    "rollup-plugin-buble": "^0.19.6",
-    "rollup-plugin-node-resolve": "^4.0.1"
+    "ol": "^5.3.0",
+    "prettier": "^2.0.5",
+    "rollup": "^2.7.2"
   },
-  "dependencies": {},
-  "peerDependencies": {},
+  "peerDependencies": {
+    "ol": "^5.3.0"
+  },
   "homepage": "https://github.com/NieuwlandGeo/SLDReader#readme",
   "directories": {
     "doc": "docs",
     "test": "test"
-  }
+  },
+  "dependencies": {}
 }
